% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract_files.R
\name{extract_files}
\alias{extract_files}
\title{Extract files matching a pattern}
\usage{
extract_files(source_dir, pattern, destination_dir = NULL,
  remove_duplicates = FALSE, unique_fun = NULL, verbose = TRUE)
}
\arguments{
\item{source_dir}{the directory to look for the files}

\item{pattern}{the pattern of files (NULL for all files) to extract}

\item{destination_dir}{the directory to which files will be copied (NULL if
files should not be copied)}

\item{remove_duplicates}{TRUE if duplicate files (based on the function supplied in unique_fun)
should be removed from the list (they will not be deleted). If destination_dir is specified,
the duplicate files will not be copied. If a file to be copied already exists in destination_dir
(as determined by unique_fun), the file will not be copied but the name of the existing file will
be included in the list of files returned.}

\item{unique_fun}{the function that will return a unique value for unique files. the most
reliable is probably tools::md5sum() (the default), but for large files this is slow
and file.size() may be adequate. In some cases basename() may also be sufficient.}

\item{verbose}{TRUE if status messages should be printed.}
}
\value{
A list of file paths. If destination_dir is specified, the list of file paths will be
  to the copied files. If not, the list of file paths will be relative to the working directory.
}
\description{
Extracts files matching a pattern from a directory, copying and/or removing
duplicate files (based on a user-supplied function) if specified. If a
destination directory is specified, files will be copied (and not
overwritten, even if filenames are identical).
}
\examples{
extract_files(getwd(), "*.*", remove_duplicates=TRUE)
}

