% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getexif.R
\name{getexif}
\alias{getexif}
\title{Get EXIF data for images files}
\usage{
getexif(filename, quiet = FALSE)
}
\arguments{
\item{filename}{A vector of filenames (like that generated by \code{list.files(..., full.names=TRUE)})}

\item{quiet}{\code{FALSE} if status updates are desired, \code{TRUE} otherwise.}
}
\value{
a data.frame of photo EXIF information
}
\description{
Checks for command length, such that windows command length policies are respected.
}
\examples{
files <- list.files(path.package("rfuncs"), recursive=TRUE, pattern="*.png", full.names=TRUE)
exifinfo <- getexif(files)
}

