% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeduplicates.R
\name{mergeduplicates}
\alias{mergeduplicates}
\title{Clean Duplicates in Dataframes}
\usage{
mergeduplicates(df, idfields, sep = "|")
}
\arguments{
\item{df}{Data frame containing duplicates}

\item{idfields}{The fields that should be unique}

\item{sep}{The separator to pass to \code{paste()} when concatenating duplicate
fields together}
}
\value{
The resulting data frame.
}
\description{
function to merge duplicate rows in a data.frame based on a set of id fields.
other fields are concatenated together using paste(x, collapse="| "). This allows
for cleaning based on all available information.
}
\examples{
df <- data.frame(x=c(1,2,3,4,1,2,3,4), 
                 y=c(1,2,3,4,1,2,3,4), 
                 label=c("one", "two", "three", "four", 
                         "five", "six", "seven", "eight"))
df
mergeduplicates(df, idfields=c("x", "y"))


}

