% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_highwaynodes_extract.R
\name{osmdata_highways}
\alias{osmdata_highways}
\title{Extract highways from Open Street Map data}
\usage{
osmdata_highways(box, apisource = osmar::osmsource_api())
}
\arguments{
\item{box}{A bounding box as generated by \code{prettymapr::searchbbox} or
\code{osmar::center_bbox}}

\item{apisource}{An OSM api source as generated by the \code{osmar::osmsource_*}
functions.}
}
\value{
An \code{osmar} object containing nodes and ways of highways.
}
\description{
Uses the \code{osmar} package to fetch OSM data and extract only highway
nodes and ways.
}
\examples{
wv <- prettymapr::searchbbox("wolfville, ns", source="google")
d2 <- osmdata_highways(wv)

}

