% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osm_highwaynodes_extract.R
\name{osmdata_nearest_nodes}
\alias{osmdata_nearest_nodes}
\title{Get the nearest OSM nodes to a point}
\usage{
osmdata_nearest_nodes(osmarobj, lon, lat, num = 1)
}
\arguments{
\item{osmarobj}{an \code{osmar} object}

\item{lon}{The longitude}

\item{lat}{The latitude}

\item{num}{The number of points to return}
}
\value{
a \code{list} containing the OSM data pertaining to the points as well
  as the computed distances
}
\description{
Finds the nearest OSM node(s) to a lat/lon coordinate.
}
\examples{
#use extract_highways to get highway data
wv <- prettymapr::searchbbox("wolfville, ns", source="google")
d2 <- osmdata_highways(wv)

#use nearest_nodes to get nearest road names to a location
nearesthwys <- osmdata_nearest_nodes(d2, -64.3557, 45.0857, num=10)
waysdata <- subset(d2, way_ids = nearesthwys$info$way_ids)
waynames <- unique(as.character(waysdata$ways$tags$v[waysdata$ways$tags$k=="name"]))
}

