% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_kml.R
\name{write_kml}
\alias{write_kml}
\title{Write Simple KML File With Placemarks}
\usage{
write_kml(lons, lats, filename, docname = "kml_document",
  names = "placemark", descriptions = "")
}
\arguments{
\item{lons}{The longitude(s)}

\item{lats}{The latitude(s)}

\item{filename}{The filename to write}

\item{docname}{The title of the document}

\item{names}{The name(s) for each placemark}

\item{descriptions}{The description(s) for each placemark}
}
\description{
Writes a simple KML File with placemarks that have names, locations, and descriptions.
}
\examples{
locs <- c("wolfville, ns", "halifax, ns", "yarmouth, ns", "sydney, ns", "amherst, ns")
lons <- c(-64.35984, -63.57532, -66.11738, -60.19422, -64.21672)
lats <- c(45.09176, 44.64886, 43.83746, 46.13679, 45.81667)

write_kml(lons, lats, filename="mykml.kml", names=locs)


}

